requirements.txt
tensorflow==2.13
numpy
matplotlib
streamlit
pillow==10.0.0
tqdm